Fix SSR for login
