+++ next.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  env: {
+    auth_domain: process.env.AUTH_DOMAIN,
+    auth_client_id: process.env.AUTH_CLIENT_ID,
+    origin: process.env.AUTH_ORIGIN
+  }
+}
