--- next.config.js.sample
@@ -1,7 +1,0 @@
-module.exports = {
-  env: {
-    auth_domain: 'church-office.eu.auth0.com',
-    auth_client_id: 'LidJYkRC9v1aU6Ue8Ey9hBpNOoLb1AgP',
-    origin: 'http://localhost:3000'
-  }
-}
