+++ package.json
@@ -22,37 +22,37 @@
   },
   "homepage": "https://github.com/berlin-church/church-office#readme",
   "dependencies": {
-    "@auth0/auth0-spa-js": "^1.7.0",
-    "@material-ui/core": "^4.9.12",
-    "@material-ui/styles": "^4.9.10",
-    "next": "^9.3.5",
-    "ra-data-json-server": "^3.4.3",
+    "@auth0/auth0-spa-js": "^1.10.0",
+    "@material-ui/core": "^4.11.0",
+    "@material-ui/styles": "^4.10.0",
+    "next": "^9.4.4",
+    "ra-data-json-server": "^3.6.2",
     "react": "^16.13.1",
-    "react-admin": "^3.4.3",
+    "react-admin": "^3.6.2",
     "react-dom": "^16.13.1"
   },
   "devDependencies": {
-    "@types/node": "^13.13.2",
-    "@types/react": "^16.9.34",
-    "@typescript-eslint/eslint-plugin": "^2.29.0",
-    "@typescript-eslint/parser": "^2.29.0",
+    "@types/node": "^13.13.12",
+    "@types/react": "^16.9.41",
+    "@typescript-eslint/eslint-plugin": "^2.34.0",
+    "@typescript-eslint/parser": "^2.34.0",
     "eslint": "^6.8.0",
     "eslint-config-standard": "^14.1.1",
     "eslint-config-standard-with-typescript": "^16.0.0",
-    "eslint-plugin-import": "^2.20.2",
-    "eslint-plugin-jest": "^23.8.1",
-    "eslint-plugin-jsx-a11y": ">= 6.2.3",
+    "eslint-plugin-import": "^2.22.0",
+    "eslint-plugin-jest": "^23.17.1",
+    "eslint-plugin-jsx-a11y": "^6.3.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.19.0",
-    "eslint-plugin-react-hooks": ">= 2.5.0",
+    "eslint-plugin-react": "^7.20.3",
+    "eslint-plugin-react-hooks": "^4.0.5",
     "eslint-plugin-standard": "^4.0.1",
     "husky": "^4.2.3",
-    "jest": "^25.4.0",
-    "lint-staged": "^10.0.8",
-    "ts-jest": "^25.2.1",
-    "tslib": "^1.11.1",
-    "typescript": "^3.8.3"
+    "jest": "^25.5.4",
+    "lint-staged": "^10.2.11",
+    "ts-jest": "^25.5.1",
+    "tslib": "^1.13.0",
+    "typescript": "^3.9.6"
   },
   "lint-staged": {
     "src/**/*.*": [
