+++ src/components/auth0-provider/index.tsx
@@ -57,7 +57,6 @@ const Auth0Provider = ({
 
   useEffect(() => {
     const initAuth0 = async () => {
-      console.debug('initAuth0')
       const auth0FromHook = await createAuth0Client({ ...initOptions })
       setAuth0(auth0FromHook)
